server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Pillow==11.2.1
Werkzeug==2.3.7
numpy==1.26.4
scipy==1.12.0
waitress==3.0.0
